requirements.txt
polars
getdaft
boto3